modify controller and services for CRUD

+++ src/api/src/main/java/com/pbkk/notificationservice/service/NotificationService.java
@@ -5,7 +5,6 @@ import java.util.List;
 import com.pbkk.notificationservice.model.Notification;
 
 public interface NotificationService {
-	List<Notification> getUserNotifications(Integer userId);
+	List<Notification> getUserNotifications(Long userId);
 	Notification createNotification(Notification notification);
-	
 }
